pyproject.toml
[tool.poetry]
name = "countrybot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
discord = "^1.0.1"
countryinfo = "^0.1.2"
regex = "^2020.11.13"
iso3166 = "^1.0.1"
country-converter = "^0.7.2"
pycountry = "^20.7.3"
pytz = "^2021.1"
iso639 = "^0.1.4"
pandas = "^1.2.2"
wbdata = "^0.3.0"
ipython = "^7.20.0"
wget = "^3.2"
a-world-of-countries = "^1.0.0"
fuzzywuzzy = "^0.18.0"
emojiflags = "^0.1.1"
replit = "^2.0.0"
dblpy = "^0.4.0"
psutil = "^5.8.0"
motor = "^2.4.0"
dnspython = "^2.1.0"





[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
